Renamed CC2CCStacked to CC2CCRelocating
