+++ src/main/java/cubicchunks/converter/lib/convert/cc2ccrelocating/CC2CCRelocatingLevelInfoConverter.java
@@ -21,25 +21,22 @@
  *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  *  THE SOFTWARE.
  */
-package cubicchunks.converter.lib.convert.cc2ccstacked;
+package cubicchunks.converter.lib.convert.cc2ccrelocating;
 
 import cubicchunks.converter.lib.Dimensions;
 import cubicchunks.converter.lib.convert.LevelInfoConverter;
 import cubicchunks.converter.lib.convert.data.CubicChunksColumnData;
 import cubicchunks.converter.lib.util.Utils;
 
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 
-public class CC2CCStackedLevelInfoConverter implements LevelInfoConverter<CubicChunksColumnData, CubicChunksColumnData> {
+public class CC2CCRelocatingLevelInfoConverter implements LevelInfoConverter<CubicChunksColumnData, CubicChunksColumnData> {
 
     private final Path srcDir;
     private final Path dstDir;
 
-    public CC2CCStackedLevelInfoConverter(Path srcDir, Path dstDir) {
+    public CC2CCRelocatingLevelInfoConverter(Path srcDir, Path dstDir) {
         this.srcDir = srcDir;
         this.dstDir = dstDir;
     }
